+++ .devcontainer/postCreateCommand.sh
@@ -26,13 +26,17 @@ npm install --global --no-fund --no-audit npm npm-check-updates shx cross-env
 echo "2 - Creating directories..."
 mkdir -p /home/node/Matterbridge /home/node/.matterbridge /home/node/.mattercert
 
-echo "3 - Setting ownership of directories... This can take a while..."
-sudo chown -R node:node . /home/node/Matterbridge /home/node/.matterbridge /home/node/.mattercert
+echo "3 - Cleaning up workspace..."
+npm run deepClean || true
 
-echo "4 - Building the package..."
+echo "4 - Setting ownership of directories..."
+ls .
+sudo chown -R node:node /home/node/Matterbridge /home/node/.matterbridge /home/node/.mattercert
+
+echo "5 - Building the package..."
 npm install --no-fund --no-audit
 npm run build
 npm link --no-fund --no-audit
 npm outdated || true
 
-echo "5 - Setup completed!"
+echo "6 - Setup completed!"