+++ CHANGELOG.md
@@ -65,7 +65,7 @@ Advantages:
 - [frontend]: Added version/tag selector to HomeInstallAddPlugins. Thanks RinDevJunior for the great idea (https://github.com/Luligu/matterbridge/pull/500).
 - [publish]: Migrate to trusted publishing / OIDC.
 - [logger]: Added parameter **--no-ansi** to disable ANSI color escape codes in all loggers. Accepts also environment variable **NO_COLOR=1** (https://no-color.org/).
-- [package]: Added sync-workspaces script for managing workspace package versions and dependencies.
+- [package]: Added version.mjs, sync-workspaces.mjs, bundle-workspaces.mjs, publish-workspaces.mjs and pack-workspaces.mjs scripts for managing monorepo packages versions and dependencies.
 
 ### Changed